--- src/components/containers/CartContainerWrapper.js
@@ -1,20 +1,0 @@
-import CartContainer from '../pages/Cart/CartContainer';
-import {connect} from "react-redux";
-
-const mapStateToProps = (state) => ({
-  products: state.cart.products,
-})
-
-const mapDispatchToProps = (dispatch) => {
-  return {
-    addToCart(product) {
-      dispatch(addToCart(product))
-    },
-    removeFromCart(product) {
-      dispatch(removeFromCart(product))
-    }
-  }
-}
-
-connect(mapStateToProps, mapDispatchToProps)(CartContainer);
-
