+++ src/components/pages/shared/TextBox.js
@@ -1,7 +1,6 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import {Link} from "react-router-dom";
-import {productPath} from "../../../helpers/routes";
 
 class TextBox extends Component {
