--- src/constants/actionTypes/ProductActionTypes.js
@@ -1,3 +1,0 @@
-export const FETCH_PRODUCT_REQUEST = 'FETCH_PRODUCT_REQUEST';
-export const FETCH_PRODUCT_SUCCESS = 'FETCH_PRODUCT_SUCCESS';
-export const FETCH_PRODUCT_ERROR = 'FETCH_PRODUCT_ERROR';
