+++ src/routes/Main.js
@@ -1,4 +1,3 @@
-import MainPage from "../components/pages/Main/MainPage";
 import CatalogContainer from "../components/containers/CatalogContainer";
 import {mainPath} from "../helpers/routes";
